Added flat toolbar-less editor.
Using current editor, added a `readOnly` to allow for a flat view of the editor.

This allows to embed the editor in any container, instead of forcing the editor to be in it's own container.

This commit could be replaced with a `viewer` instead, but would duplicate alot of the editor.