Moved review/approve button to left pannel.
The review button used to be only accessible through general information.

The button was moved at the bottom of the sections (could use some more thinking?)

Also fixed an issue that allowed reviewers to approve without the report being ready to review.

The app does not seem to check for this in backend (maybe an issue already?)